+++ NEWS.md
@@ -11,7 +11,7 @@ Version 20.0-dev (master)
 * Install user manual to /usr/share/doc to better conform to the GNU
   standard directories and because Firefox and Chromium on Ubuntu 21.10
   have no permissions to open files in /usr/share/help.
-* Migrated Android version from Qt 5.15 to Qt 6.2.
+* Support for compiling Android app with Qt >=6.2.
 * Faster move generation in multi-player game variants.
 * Install application and mimetype bitmap icons for multiple sizes.
 * Show error message if launching web browser failed.